+++ lib/ui.dart
@@ -59,6 +59,8 @@ part 'src/ui/vertical_panel.dart';
 part 'src/ui/horizontal_panel.dart';
 part 'src/ui/dock_layout_panel.dart';
 part 'src/ui/layout_command.dart';
+part 'src/ui/layout_panel.dart';
+part 'src/ui/root_layout_panel.dart';
 
 part 'src/ui/widget_collection.dart';